Add storage engine interface, move issuers to be stored in engine
This commit introduces the concept of a storage engine in DMV, as well
as a reference in-memory implementation of issuer storage. Leaving
subject token issuers as a Fosite-specific configuration item yielded
a number of issues, and introducing a new set of storage engine
interfaces allows us to separate data persistence from OAuth service
implementation using either in-memory static data or, in the future, a
database.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>
Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>