Use a value receiver for ErrorIssuerNotFound
errors.Is does shallow comparison of error values, meaning for
pointer values it expects the two errors to be the same pointer. This
causes problems when using errors.Is with error structs. This commit
updates ErrorIssuerNotFound to implement the error interface using a
value receiver instead, so that errors.Is works without requiring an
explicit implementation of Is on the error.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>