Leverage SSH agent forwarding instead of SSH_AUTH_SOCK mounting
SSH agent forwarding is generally a cleaner way to handle getting the
agent into the dev container rather than fiddling with ControlMaster
settings. This commit adds the sshd feature to the dev container and
updates the README for instructions on how to use it.

This commit also changes the casing of Dev Container because I didn't
really like it.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>