Added UserInfo in memory store
This adds the UserInfo storage interface and an in-memory
implementation of that. This interface is intended to allow the
TokenEndpoint handler to populate user info for subject tokens, and
for a future UserInfo endpoint to retreive that information.

* Fetch and store UserInfo while handling token exchange request

This adds the UserInfoInterface and an associated Provider to the
Oauth2Configurator interface so we can pass the dependency down to the
handler. Once in the token exchange handler, the issuer and subject
token are passed to the userinfo service to retreive the information
and to store it in the storage backend.

* Use userinfo ID as subject

This allows for us to do lookups in both directions with both the
subject and the exchanged token

Signed-off-by: Adam Mohammed <admohammed@equinix.com>