Add management API (#33)
* Move internal packages under internal/

In order to provide a good management API, we should start with only
exposing the parts of the public API that should be public. This
commit moves everything that isn't the public API under internal/ in
line with other Infratographer services.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Move api/v1 package to internal/types

api/v1 is to be reserved for HTTP API types. This commit moves the
internal representation of issuers to an internal package so the
exported interface represents what client libraries can use.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add OpenAPI-driven management API

This commit introduces an OpenAPI 3.0-driven management API for
identity-manager-sts, including:

* Make targets for code generation and dependency fetching
* Request validation against the provided OpenAPI spec
* Strict handling of requests using Go structs for better type safety

Current operations supported include creation and fetching of
issuers.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Move ErrorIssuerNotFound to types and make it a variable

ErrorIssuerNotFound is an internal error type, not an API
response. This commit moves it into the types package and makes it a
variable so it's easier to make comparisons.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add better error handling to management API

The error handling in the management API (as well as the generated
oapi-codegen code) left a bit to be desired. This commit adds better
error handling, including consistent response formats following
validation errors and better distinction between 4xx errors and
generic 500s.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Use types.IssuerService for API handler instead of storage.Engine

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Rename IssuerService methods to include Issuer in method names

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Remove spurious debugging

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add support for deleting issuers

This commit adds support for DELETE /api/v1/issuers/{id}.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add support for issuer updates and PATCH /api/v1/issuers/{id}

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add testingx package, update tests to use it

Packages that have tests generally follow the same table testing
format, including some that use generics already. This commit factors
all of that into a common testingx package and updates tests to use
it.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add tests for MemoryIssuer, fix bug in UpdateIssuer

This commit adds tests for MemoryIssuer and also fixes a bug where
updates to JWKSURI would fail.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add tenant ID to issuer

Issuers should be a sub-resource of tenants. This commit makes it
so.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add API handler tests

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add context to testingx CheckFn

There are situations in which using the test context is desirable when
checking test results, such as when testing side effectful
operations. This commit adds context.Context to the CheckFn type
signature in testingx.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add optional setup/cleanup functions to testingx test cases

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add support for transactions to storage engine

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add storage middleware for Gin, refactor error responses

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Remove extra response codes from spec + generated code

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Add uniqueness constraint to issuer URI

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Provide a successful delete response even when issuers do not exist

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Rename memory issuer service tests to match method names

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Remove unused oapi-codegen config file

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Ensure cleanup of test DB

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Make sure testingx tests close over ctx

The race detector found a race condition. This fixes the race
condition.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

---------

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>