Add helm-chart
Deploying IM-STS requires a bit of configuration to get a working
deployment. This helm-chart makes sure that we create the right k8s
resources. It depends on having secrets in places for the
`oauth-secret` and for the private keys used to create tokens. If
using the ingress, it also requires tls configuration to serve HTTPS.

Signed-off-by: Adam Mohammed <admohammed@equinix.com>