Added UserInfo in memory store (#32)
* Added UserInfo in memory store

This adds the UserInfo storage interface and an in-memory
implementation of that. This interface is intended to allow the
TokenEndpoint handler to populate user info for subject tokens, and
for a future UserInfo endpoint to retreive that information.

* Fetch and store UserInfo while handling token exchange request

This adds the UserInfoInterface and an associated Provider to the
Oauth2Configurator interface so we can pass the dependency down to the
handler. Once in the token exchange handler, the issuer and subject
token are passed to the userinfo service to retreive the information
and to store it in the storage backend.

* Use userinfo ID as subject

This allows for us to do lookups in both directions with both the
subject and the exchanged token

Signed-off-by: Adam Mohammed <admohammed@equinix.com>

* Add tests with the testingx internal package helper

Signed-off-by: Adam Mohammed <admohammed@equinix.com>

* Add /userinfo endpoint

This adds the `/userinfo` endpoint, which when hit with a token minted
by the STS, returns the user info as fetched from the origin IdP.

Signed-off-by: Adam Mohammed <admohammed@equinix.com>

* Add tests for LookupUserInfoByID

Signed-off-by: Adam Mohammed <admohammed@equinix.com>

* Remove extra variable

Signed-off-by: Adam Mohammed <admohammed@equinix.com>

* Improve UserInfo struct docstring

Signed-off-by: Adam Mohammed <admohammed@equinix.com>

* Bump to go.hollow.sh/toolbox 0.5.0

Signed-off-by: Adam Mohammed <admohammed@equinix.com>

* Use constants for column names and omit empty userinfo fields

Signed-off-by: Adam Mohammed <admohammed@equinix.com>

* Add more validation around storing userinfo

Signed-off-by: Adam Mohammed <admohammed@equinix.com>

* Compress query formatting with explicit format indexes

Signed-off-by: Adam Mohammed <admohammed@equinix.com>

* Added transaction manager interface and used that for userinfo

Signed-off-by: Adam Mohammed <admohammed@equinix.com>

---------

Signed-off-by: Adam Mohammed <admohammed@equinix.com>