Change fositex.NewOAuth2Provider to behave like fosite Compose
For other oauth2 token handler types, fosite has a compose package
that allows passing a config, store, and strategy to the factories
for each handler. The compose function also handles attaching the
resulting types to the right parts of the `fosite.Config` so that they
are registered in the correct places.

Signed-off-by: Adam Mohammed <admohammed@equinix.com>