Add OAuthClientManager and implementation
In order to have a client credentials flow, we need to be able to
store and retrieve oauth clients. This PR adds the storage interface
and updates the spec to have the relevant endpoints. In order to get
the API types, I ran `make generate` and added stub endpoints so that
the application could build.

I will add the API and credential flow implementations in a separate
PR because both come with a bunch of changes and I'm trying to limit
the size of this PR.

Signed-off-by: Adam Mohammed <admohammed@equinix.com>