Remove in-memory storage engine, refactor config logic (#94)
* Remove in-memory storage engine, refactor config logic

By convention, Infratographer CRDB configs should be a top-level
config item. In reconfiguring this, a number of other things shook out
including no longer needing (or necessarily even wanting) a separate
in-memory storage engine. The dev container supports a database now
and we can just create CRDB test servers as needed for testing. This
commit refactors the storage package to take a crdbx config directly
and defines migrations, tracing, and database seeding as options on
the storage engine.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Shut down database on engine.Shutdown()

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

* Remove Engine.Shutdown entirely

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>

---------

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>