Fetch UserInfo using endpoint defined by OIDC Discovery
Currently, identity-api fetches UserInfo for subjects by appending
/userinfo directly to the issuer URL. This is incorrect according to
the OIDC Discovery spec, which specifies that a client should make a
request to /.well-known/openid-configuration and then use the
"userinfo_endpoint" value in that response as the UserInfo
endpoint. This commit updates FetchUserInfoFromIssuer to be in line
with the OIDC Discovery spec's expected behavior.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>