Make subject ID generation deterministic
There are a number of use cases where it is advantageous to make the
value of the "sub" claim in identity-api access tokens
deterministic. This commit updates the token format design document
with a deterministic algorithm for subject ID generation based on the
subject token "iss" and "sub" claim values, and additionally updates
identity-api to implement this algorithm and use it during token
exchange.

Fixes #166.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>